Update README.md
This project demonstrates the prediction of customer churn using a dataset of bank customers. The goal is to identify customers who are likely to leave the bank based on various features such as Credit Score, Age, Geography, and Balance. The dataset is processed, and a Support Vector Machine (SVM) classifier is used for prediction. Random under-sampling and over-sampling are also applied to handle class imbalance.